improve peeking of unowned cards
Replace the piece directive with custom one which always loads the
colored image but makes it gray with CSS if needed. When the user
press the image, it gets either colored or replaced with video of
the animated version. Also, if the piece is animated and the user
owns it, the original version will be displayed instead.